Remove the ability to set markup style dynamically.
Since the markup is inherently tied to the text content of a segment, it makes more sense to just provide two constructors depending on the type of markup used rather than permitting the markup style to be changed later. Since the text cannot be changed later, allowing markup style changes could also result in unexpected formatting when displayed in i3bar.