Update events to include relative_x/y and width/height.
This was recently added to i3: https://github.com/i3/i3/issues/2767. Also renames X,Y to be relative to the segment, so that X,Y,Width,Height make sense, and provides the previous X,Y as ScreenX,ScreenY to make it more obvious that they are relative to the root window.