Remove 'Module' interfaces, use *Module instead.
The reason for the interfaces was because of the embedded base.Base, which exposed functions like Output, Error, etc. directly on the module. The interfaces served to restrict the visible methods to a subset, like Stream and Click. Now that base.Base is no longer, we can also dispense with the Module interfaces and just return *Module types since all exported functions are controlled by the module. This makes modules more go-like (see, e.g. *time.Ticker, *testing.T, *http.Request) by returning a pointer, and less java-like (interfaces).