volume: Add PulseAudio (D-Bus) implementation.
This volume module implementation connects to a running PulseAudio
server using its D-Bus interface. The PulseAudio server needs to have
D-Bus enabled - load module-dbus-protocol in your default.pa, or load
it at runtime with pactl.

To use the PulseAudio volume module, use volume.DefaultSink() to create
a module that controls whatever the default sink is, even if it changes
at runtime, or volume.Sink('<PA sink name>') to create a module that
controls a specified sink. For Bluetooth headphone users, use
volume.DefaultSink(), because it will control your Bluetooth headphones
if they are connected, or your default sound card if they are not.