First class error outputs.
Adds support for an error property on segment, and adds special handling for right-clicks on error segments. Default implementation simply calls `i3-nagbar -m` with the error text, but the triggering event is provided to the error handling to allow error UI to be aligned with the segment triggering it.

First class error outputs also make the AssertError tests far more robust by allowing them to actually look for errors rather than using the previous heuristic of urgent + short_text. Since errors are now a part of the segment, the segment text can default to 'Error' as well, which means a really long error will not force all modules into compact mode.

Fixes #29.