Use *Segment instead of Segment, remove unneeded wrapping struct
The only reason the struct existed was to give Segment{} pointer semantics by wrapping a pointer to the actual data. Exported pointer types are present in stdlib, so using exported pointer to Segment is fine.
Exporting the Segment type but using it as a pointer also has the advantage of allowing a zero-value &Segment{}, while in the previous case, Segment{} would panic because of the nil reference to data.
It makes the potential race condition more explicit as well, since we're sending a pointer over the output channel, but in most cases that pointer is never accessed by the sender again, so it's a minor concern. This race was already present, but harder to track down because of the struct {*data} wrapper.