Convert bar.Module from Stream() <-chan bar.Output to Stream(bar.Sink)
This is a big refactor, but on the whole it results in much cleaner code. The original idea of barista was to play to one of the few strengths of go, namely channels. But after implementing several modules, a step back revealed that most modules were getting more complex (not less) because of the use of channels. Now channels are mostly restricted to the core and testing utilities. Most modules can now be implemented in less code, and modules do not need to care about channel management (esp. write after close), or about runaway goroutines since they never start any.

Fixes #37