Accept -finelog in addition to --finelog.
Although --finelog is more appropriately in-line with unix tools, go has decided to eschew single letter flags and instead use a single hyphen with named flags. So it can be confusing (especially when running test) to provide some arguments with a single hyphen and then `--finelog=foo'. So, handle both -finelog and --finelog.