Rename Pango() to String()
Pango() is a relic from a time when pango was implemented with a lot of interfaces. Now that everything is a concrete *Node, there's no need to have a method named 'Pango()' (since there is no longer an interface for it to be a part of), and calling it String() makes more sense.