Only update CodeClimate if test reporter id is available
Since codecov.io is in travis.yml, running travis.sh locally does not affect it. But since code climate is in travis.sh, running it locally will probably cause problems with the missing env var.
Disabling code climate when the env var is not available is a reasonable solution for now. Todo: move code climate to travis.yml as well.