Add barista/core to provide a feature-rich wrapper around Module
Also adds a ModuleSet that can be useful to host multiple modules and interact with them in the abstract, without needing to manually handle the click-to-restart functionality.
This paves the way for simplifying group, as well as more robust tests that simply use a core/ModuleSet instead of a full bar running on mockio.