Reduce flakiness with output function for all modules
By ensuring the subscription happens before the output, we reduce the chances that the call to Output(...) in tests happens before the .Next() in the module. Currently there is a small window during which an update will be discarded, failing the test. This can happen if the test code calls Output(...) before the module has a chance to enter the next for loop iteration and subscribe to the output function updates.