Revert 2952552, fixes deadlock in group
Concurrency is hard.

This fixes the deadlock introduced in the original commit, which can be observed in the following scenario:
- group acquires the RLock() via grouper.Lock()
- external source attempts to write, e.g. changing the current "mode"
- external source is in the rwmutex queue with a write
- group generates the output, which causes the output function to call controller.Current()
- controller.Current() tries to acquire another RLock()

The previous commit made the incorrect assumption that if RLock is held, subsequent RLock() will always succeed. This is not the case if a Lock() is in the queue (due to mutation).
So now grouper holds one RLock() and is waiting for the second RLock(), while the external source is waiting on Lock(). This is a deadlock.