Move SubscribeTo to Signaller, rename to Source
Value/ErrorValue were the only real users of SubscribeTo, so moving to Source allows them to provide a Subscribe method directly. Cleaner (avoids importing notifier at all call sites), and makes Source more powerful.

The other uses of SubscribeTo will be migrated to <-chan struct{} exported struct members, making it clearer that the channel will not be closed. This is a pattern used by stdlib time.Ticker, so it should be fine for use here.