Add properties watcher to dbus
WatchProperties can watch a specific interface on a specific object, obtained using a service, and will notify with old and new values for each property change.
Also presents a snapshot of the current properties if needed, as well as signal handling in case certain properties do not trigger PropertiesChanged.

Fixes #39