Move base/sink to sink, use Output to make it more extensible
Using bar.Output as the type allows extensions that can be detected at runtime. The current implementation immediately converts all outputs to segments, losing any extensions in the process.
This has the small downside of holding a ref to the whole output instead of just the generated segments, which means we now require that Outputs not be used after being sent on a sink.

Works towards #49