Add RefresherModule to allow middle-click to update data
This allows on-demand updates of data that is infrequently fetched (calendar, weather, email, etc.) by middle clicking on bar output to refresh data.
Error outputs are handled differently, if a refreshable module sends an error output and does not finish, then a restartable click on the error will also refresh the data.