Increase timeout for expected events to 10s
Under extreme system load, it's possible that an action in a different goroutine ends up taking more than 1s for the combined start + perform action + signal to test. This is especially true for actions where the effect is measured across several hops (e.g. a module update goes through a core Module, then a core ModuleSet, then the test bar instance, and finally to the test.)