Track machine time zone changes in clock.Local
Go pins time.Local at the start of the binary (golang/go#28020), but because the bar can run for days on end even as the machine moves around the world, we need to track the machine timezone for local clock modules.
This changes clock.Module to handle a nil timezone as "true" machine-local time zone, updating the output any time the timezone changes locally.