Expand unit formatting capabilities
Adds format.Unit, format.SI to format various unit.* values automatically. This can be used by other parts of the bar to provide sensible default output when displaying these values. The width parameter allows for fixed output sizes (within reason) so that the bar does not shift when the value changes (e.g. from 5 to 15).

The unit and value are exported separately to allow further customisation when using pango (e.g. to show units in a smaller or different font, or to allow clicking on the unit to change the unit used).