Retry flaky tests thrice before failing the suite
This should significantly reduce the CI resource usage, since currently a flaky test requires a full rebuild and test. With this change, only the specific test that fails needs to be re-run, which should be a few seconds of extra time.

It also prints the failures before eventually passing, so historical logs will now start to provide a better indication of which tests are flaky and how much.