Fix apt-get state in actions
Why this doesn't work out of the box, I have no idea. I expected whatever worker/drone runs actions would be able to stay up-to-date with the apt repositories, but that appears to not be the case. Instead, I need to manually run apt-get update each time because apparently there's a drift between the cached packages index and what's actually available out in the world.

This does not leave me with a good impression of GitHub actions, but the lack of an external dependency is _probably_ still worth this hack.